feat: 예약 확정 요청을 위한 Command 인터페이스 추가 및 V1, V2 구현 분리
- 버전별로 Command가 달라 공통 인터페이스 적용 후 전략 패턴에서 적용
