+++ reservation-service/src/main/java/com/oneul_tanda/reservation_service/reservation/application/command/ConfirmReservationCommandV1.java
@@ -1,0 +1,50 @@
+package com.oneul_tanda.reservation_service.reservation.application.command;
+
+import com.oneul_tanda.reservation_service.reservation.domain.entity.vo.Gender;
+import com.oneul_tanda.reservation_service.reservation.presentation.dto.request.update.ConfirmReservationRequestDto;
+
+import java.util.List;
+
+import java.util.UUID;
+
+public record ConfirmReservationCommandV1(
+        UUID userId,
+        UUID reservationId,
+        List<ConfirmTicketCommand> tickets
+) implements ConfirmReservationCommand {
+    public static ConfirmReservationCommandV1 of(UUID userId, UUID reservationId, ConfirmReservationRequestDto dto) {
+        return new ConfirmReservationCommandV1(
+                userId,
+                reservationId,
+                dto.tickets().stream()
+                        .map(ConfirmTicketCommand::from)
+                        .toList()
+        );
+    }
+
+
+
+    public record ConfirmTicketCommand(
+            UUID ticketId,
+            ConfirmPassengerCommand passenger
+    ) {
+        public static ConfirmTicketCommand from(ConfirmReservationRequestDto.ConfirmTicketDto dto) {
+            return new ConfirmTicketCommand(dto.ticketId(), ConfirmPassengerCommand.from(dto.passenger()));
+        }
+    }
+
+
+
+    public record ConfirmPassengerCommand(
+            String name,
+            String birth,
+            Gender gender,
+            String passportNumber
+    ) {
+        public static ConfirmPassengerCommand from(ConfirmReservationRequestDto.ConfirmPassengerDto dto) {
+            return new ConfirmPassengerCommand(dto.name(), dto.birth(), dto.gender(), dto.passportNumber());
+        }
+    }
+}
+
+
